requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
audioop-lts==0.2.1
bcrypt==4.2.1
blinker==1.9.0
buildozer==1.5.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
distlib==0.3.9
docutils==0.21.2
ecdsa==0.19.0
fastapi==0.115.8
filelock==3.17.0
filetype==1.2.0
Flask==3.1.0
fpdf==1.7.2
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
googletrans==4.0.0rc1
greenlet==3.1.1
gTTS==2.5.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Kivy==2.3.1
Kivy-Garden==0.1.5
kivy_deps.angle==0.4.0
kivy_deps.glew==0.3.1
kivy_deps.sdl2==0.8.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
narwhals==1.27.1
numpy==2.2.3
openai==0.28.1
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
pluggy==1.5.0
propcache==0.2.1
protobuf==5.29.3
ptyprocess==0.7.0
pyarrow==19.0.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pypiwin32==223
pytest==8.3.4
python-dateutil==2.9.0.post0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.1
pywin32==308
referencing==0.36.2
requests==2.32.3
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
sh==2.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SpeechRecognition==3.14.1
SQLAlchemy==2.0.38
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.45.3
streamlit==1.42.1
streamlit-calendar==1.2.1
stripe==11.5.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.2
watchdog==6.0.0
Werkzeug==3.1.3
yarl==1.18.3
